v3.0.0 - Adding init --force

+++ src/DbBroker.Cli/Commands/Init/InitCommand.cs
@@ -5,8 +5,8 @@ namespace DbBroker.Cli.Commands.Init;
 public class InitCommand
 {
 
-    private const string DbBrokerConfigFileContents =
-    """
+  private const string DbBrokerConfigFileContents =
+  """
 {
   "$schema": "https://raw.githubusercontent.com/diegosiao/DBBroker/refs/heads/v3.0.0/docs/json-schema/dbbroker.config.schema.json",
   "contexts": [
@@ -21,27 +21,33 @@ public class InitCommand
 }
 """;
 
-    internal static int Execute(InitOptions opts)
-    {
-        var dbbrokerconfigFilePath = Path.Join(Directory.GetCurrentDirectory(), "dbbroker.config.json");
+  internal static int Execute(InitOptions opts)
+  {
+    var dbbrokerconfigFilePath = Path.Join(Directory.GetCurrentDirectory(), "dbbroker.config.json");
 
-        try
-        {
-            File.WriteAllText(
-                dbbrokerconfigFilePath,
-                DbBrokerConfigFileContents
-                    .Replace("$NAMESPACE$", opts.Namespace)
-                    .Replace("$CONNECTIONSTRING$", opts.ConnectionString)
-                    .Replace("$PROVIDER$", opts.Provider.ToString()));
-            
-            $"Configuration file created successfully: '{dbbrokerconfigFilePath}'.".Success();
-            return ExitCodes.SUCCESS;
-        }
-        catch (Exception ex)
-        {
-            Console.Error.WriteLine(ex.Message);
-            Console.Error.WriteLine(ex.StackTrace);
-            return ExitCodes.CONFIG_FILE_INIT_ERROR;
-        }
+    try
+    {
+      if (File.Exists(dbbrokerconfigFilePath) && !opts.Force)
+      {
+        "The 'dbbroker.config.json' file already exists. Use --force to replace it.".Error();
+        return ExitCodes.CONFIG_FILE_INIT_ERROR;
+      }
+
+      File.WriteAllText(
+          dbbrokerconfigFilePath,
+          DbBrokerConfigFileContents
+              .Replace("$NAMESPACE$", opts.Namespace)
+              .Replace("$CONNECTIONSTRING$", opts.ConnectionString)
+              .Replace("$PROVIDER$", opts.Provider.ToString()));
+
+      $"Configuration file created successfully: '{dbbrokerconfigFilePath}'.".Success();
+      return ExitCodes.SUCCESS;
     }
+    catch (Exception ex)
+    {
+      Console.Error.WriteLine(ex.Message);
+      Console.Error.WriteLine(ex.StackTrace);
+      return ExitCodes.CONFIG_FILE_INIT_ERROR;
+    }
+  }
 }
